Move import cli functionality to its own modules
