+++ dae/dae/import_tools/tests/test_import_configs.py
@@ -7,7 +7,7 @@ import yaml
 import pytest
 import fsspec
 
-from dae.import_tools import import_tools
+from dae.import_tools import import_tools, cli
 from dae.configuration.gpf_config_parser import GPFConfigParser
 
 
@@ -31,7 +31,7 @@ def test_parquet_files_are_generated(tmpdir, gpf_instance_2019, config_dir,
     project = import_tools.ImportProject.build_from_config(
         import_config, input_dir)
 
-    import_tools.run_with_project(project)
+    cli.run_with_project(project)
 
     files = os.listdir(tmpdir)
     assert len(files) != 0
@@ -60,7 +60,7 @@ def test_import_with_add_chrom_prefix(tmpdir, gpf_instance_grch38, mocker,
                         return_value=gpf_instance_grch38)
     project = import_tools.ImportProject.build_from_config(
         import_config, input_dir)
-    import_tools.run_with_project(project)
+    cli.run_with_project(project)
 
     files = os.listdir(tmpdir)
     assert len(files) != 0
