+++ dae/dae/inmemory_storage/tests/test_inmemory_genotype_storage.py
@@ -3,7 +3,7 @@ import re
 
 import pytest
 
-from dae.import_tools.import_tools import run_with_project
+from dae.import_tools.cli import run_with_project
 from dae.inmemory_storage.inmemory_genotype_storage import \
     InmemoryGenotypeStorage
