+++ dae/tests/integration/schema1/test_cnv_variants.py
@@ -9,7 +9,8 @@ from dae.variants_loaders.cnv.loader import CNVLoader
 from dae.pedigrees.loader import FamiliesLoader
 from dae.variants_loaders.raw.loader import AnnotationPipelineDecorator
 from dae.inmemory_storage.raw_variants import RawMemoryVariants
-from dae.import_tools.import_tools import ImportProject, run_with_project
+from dae.import_tools.import_tools import ImportProject
+from dae.import_tools.cli import run_with_project
 
 from dae.configuration.gpf_config_parser import FrozenBox
 from dae.utils.regions import Region
